#318: hide utility sub-classes of `Submission`
