+++ onlinejudge/service/yukicoder.py
@@ -350,7 +350,7 @@ class YukicoderProblem(onlinejudge.type.Problem):
         if 'submissions' in resp.url:
             # example: https://yukicoder.me/submissions/314087
             log.success('success: result: %s', resp.url)
-            return onlinejudge.type.DummySubmission(resp.url)
+            return utils.DummySubmission(resp.url, problem=self)
         else:
             log.failure('failure')
             log.debug('redirected to %s', resp.url)
